refactor(ws): move WebSocketRoute/Router/EndpointInfo to reinhardt-core; add UnifiedRouter::websocket()
