+++ crates/reinhardt-websockets/Cargo.toml
@@ -30,6 +30,7 @@ anyhow = { workspace = true }
 bytes = { workspace = true }
 once_cell = "1.19"
 inventory = { workspace = true }
+reinhardt-core = { workspace = true }
 chrono = { version = "0.4", features = ["serde"] }
 
 # Dependency injection (optional)
